requirements.txt
fastapi
uvicorn
newspaper3k
google-generativeai